requirements.txt
diffusers==0.21.4
transformers
einops
omegaconf
ftfy
tensorboard
modelcards
click
pandas
einops
loralib
regex
tqdm
pytorch_lightning
